refactor: move nobuild command, make nobuild work on rotated rooms
