+++ src/plugins/commands/spawn-object-command.plugin.ts
@@ -1,71 +1,14 @@
 import { commandActionHandler } from '@engine/action';
+import { findObject } from '@engine/config';
+import { Position } from '@engine/world';
+import { RoomType } from '@plugins/skills/construction/con-constants';
+import { ROOM_CONFIG } from '@plugins/skills/construction/data';
+import { Room } from '@plugins/skills/construction/house';
+import { getCurrentRoom } from '@plugins/skills/construction/util';
+import { rotateChunkCoordinate } from '@plugins/skills/construction/util/rotations';
 import { LandscapeObject } from '@runejs/filestore';
 
-// obj ids
-// 13098 is light wood wall
-// 13099 is light wood window
 
-
-// 13103 is light wood door
-// 13106 is ???
-// 13109 is dark wood door
-// 13110 is dark wood door
-// 13111 is dark wood wall
-// 13112 is dark wood window
-// 13113 is ???
-// 13114 is ???
-// 13115 is ???
-// 13116 is white wall
-// 13117 is white window
-
-type RoomType = 'study';
-
-type RoomObject = {
-    key: string;
-    placeholderId: number;
-    x: number;
-    y: number;
-    orientation: number;
-    objectType: number;
-}
-
-const study = {
-    objects: [
-        { key: 'wall_space_1', x: 0, y: 1, placeholderId: 15423, objectType: 4, orientation: 0 }, // wall space
-        { key: 'wall_space_2',  x: 0, y: 6, placeholderId: 15423, objectType: 4, orientation: 0 }, // wall space
-        { key: 'wall_space_3',  x: 6, y: 7, placeholderId: 15423, objectType: 4, orientation: 0 }, // wall space
-        { key: 'wall_space_4',  x: 6, y: 0, placeholderId: 15423, objectType: 4, orientation: 0 }, // wall space
-
-        { key: 'globe', x: 1, y: 1, placeholderId: 15421, objectType: 10, orientation: 0 }, // globe
-        { key: 'bookcase_1', x: 0, y: 3, placeholderId: 15425, objectType: 10, orientation: 0 }, // bookcase
-        { key: 'bookcase_2', x: 0, y: 4, placeholderId: 15425, objectType: 10, orientation: 0 }, // bookcase
-        { key: 'telescope', x: 0, y: 5, placeholderId: 15424, objectType: 10, orientation: 0 }, // telescope
-        { key: 'crystal_ball', x: 5, y: 5, placeholderId: 15422, objectType: 10, orientation: 0 }, // crystal ball
-        { key: 'lectern', x: 5, y: 2, placeholderId: 15420, objectType: 10, orientation: 0 }, // lectern
-    ]
-}
-
-const stephensStudy = {
-    'bookcase_1': '13100',
-    'bookcase_2': '13100',
-}
-
-const noBuildModeAction: commandActionHandler = ({ player, args }) => {
-    const room = study;//roomObjects[args.room as RoomType];
-
-    for (const roomObj of room.objects) {
-        const landscapeObject: LandscapeObject = {
-            objectId: roomObj.placeholderId,
-            x: player.position.x + roomObj.x,
-            y: player.position.y + roomObj.y,
-            level: player.position.level,
-            type: roomObj.objectType,
-            orientation: roomObj.orientation
-        }
-
-        player.outgoingPackets.removeLocationObject(landscapeObject, player.position.add(roomObj.x, roomObj.y));
-    }
-};
 
 const spawnObjectAction: commandActionHandler = ({ player, args }) => {
     const landscapeObject: LandscapeObject = {
@@ -101,11 +44,6 @@ const despawnObjectAction: commandActionHandler = ({ player, args }) => {
 export default {
     pluginId: 'rs:spawn_object_command',
     hooks: [
-        {
-            type: 'player_command',
-            commands: [ 'nobuild' ],
-            handler: noBuildModeAction
-        },
         {
             type: 'player_command',
             commands: [ 'wall' ],
